+++ src/components/Overview/Choices/Choice/index.js
