+++ src/components/Overview/Explanation/Sentence/index.js
