+++ src/styles/components/inference.scss
@@ -1,4 +1,4 @@
-.details {
+.inference {
   height: 100%;
   width: 100%;
   display: flex;
