--- src/styles/components/screen-switcher.scss
@@ -1,11 +1,0 @@
-.screen-switcher {
-  position: absolute;
-  top: 20px;
-  right: 20px;
-
-  &--in-box {
-    position: relative;
-    float: right;
-    top: -5px;
-  }
-}
