+++ src/styles/components/toggler.scss
@@ -1,0 +1,5 @@
+.toggler {
+  position: relative;
+  float: right;
+  top: -5px;
+}
